Д36/2: шесть классов дефектов clang-tidy, и у каждого свой примитив вместо N правок
1. UNCHECKED OPTIONAL (43 места вне тестов). Крупнейший узел — очереди Vulkan: семь
   мест разворачивали std::optional индексов семейств, три через .value() внутри
   КОНСТРУКТОРА (std::bad_optional_access без обработчика), два через * (UB), а
   CreateDevice раздавал .value_or( -1 ) как индекс семейства и продолжал. Починено не
   семь раз, а один: VulkanPhysicalDevice::CreateDevice ОТКАЗЫВАЕТ по имени, называя
   какие семейства драйвер сообщил, и после этого геттеры возвращают uint32_t — тип
   несёт гарантию. Второй узел — материалы: MaterialData::InstanceParentId() отвечает
   на "инстанс ли это" и "чей ребёнок" ОДНИМ вызовом (было: IsInstance() и разыменование
   через ВТОРОЙ вызов Data(), шесть раз). Остальное — одна форма: optional читается
   ОДИН раз в именованную привязку, и проверка с использованием говорят об одном объекте.
   Удалён мёртвый PresentFamily (не присваивался нигде) и мёртвый класс
   VulkanRenderCommandBuffer целиком (ни одного пользователя; там и жил `cond ? a : a`).

2. MEMCMP ПО ТИПАМ БЕЗ УНИКАЛЬНОГО ПРЕДСТАВЛЕНИЯ (3). Бриф называет причиной
   выравнивание — для CloudTypeShape это НЕ так: в нём одни float, а у float нет
   уникального представления при любой раскладке. -0.0f и 0.0f — одно значение и два
   набора байт, то есть небо перепекалось без изменений; два разных NaN с равными
   битами читались как одна форма. У типа уже был defaulted operator==. Vertex2D его
   получил. В LVK.hpp (vendored lightweightvk) padding РЕАЛЬНЫЙ: три uint32_t перед
   uintptr_t, четыре байта, которые никто не инициализирует.

3. EXCEPTION ESCAPE. Восемь деструкторов ЛОГИРУЮТ, то есть выделяют и форматируют:
   строка, чья работа — сделать сбой видимым, была самой вероятной причиной исчезнуть
   без слова. Common/Core/DestructorGuard.hpp: function-try-block + fputs/fflush,
   которые не выделяют и не бросают. Tools/Shared/ToolMain.hpp: тринадцать main'ов
   инструментов, у которых НЕТ обработчика; движок в этот заголовок не входит
   намеренно — EntryPoint.hpp ставит set_terminate с бэктрейсом ЧЕРЕЗ место броска,
   что строго больше, чем может сказать catch. Три noexcept в AssetHandle.hpp строили
   std::string: подпись обещала гарантию, которой тело не даёт.

4. IMPLICIT WIDENING (14 вне тестов). 32-битное произведение как 64-битное смещение:
   переполнение уже произошло к моменту, когда широкий тип его увидит.

5. INCORRECT ROUNDINGS (14). Common/Core/Math/Rounding.hpp — RoundToNearest (std::lround,
   одна инструкция fcvtas, верен для отрицательных) и QuantiseUnitToByte (клампит сам).
   Десять из четырнадцати мест клампили до неотрицательного и были верны; четыре нет, и
   на месте они выглядели одинаково. Замена доказана БИТ В БИТ на 10 000 001 выборке
   единичного интервала — ни один испечённый воксель не сдвинется.

6. МЕЛКИЕ. Присваивание в условии if (UICanvasRenderer2D); пустой catch в
   DesertHeaderTool заменён на std::from_chars, который сообщает "это не число"
   возвращаемым значением — вывод инструмента совпал байт в байт; std::strerror
   (общий статический буфер) -> Common::ErrnoText через strerror_r, 6 мест.

ГЕЙТ (закрывает все шесть сразу): scripts/CI/TidyRegister.txt — 53 файла x 7 проверок,
прогоняется из CheckTidy.sh вторым шагом ПО УМОЛЧАНИЮ, так что CI подхватывает его без
правки workflow. Перепись по строкам, а не по числу: перечислены файлы, а
Desert/Tests/Tools/TidyRegister держит восемь ИМЕНОВАННЫХ строк — по одной на класс, в
том файле, где класс был найден. Суита Desert/Tests/Common/Rounding краснеет на
возвращённой форме (отрицательные значения) и заодно меряет, что положительные не
двигаются.

МУТАЦИИ: возврат `+ 0.5` в Profiler.cpp красит --register; он же красит
Rounding.NegativeValuesRoundAwayFromZero на -0.6/-1.5/-2.5/-2.7.

+++ Desert/Desert/Source/Engine/Assets/CloudNoiseVolumeSheet.cpp
@@ -112,7 +112,8 @@ namespace Desert::Assets
                      ( static_cast<size_t>( originY + y ) * sheet.Width + originX ) * kBytesPerVoxel;
 
                 std::copy( volume.Voxels.begin() + static_cast<std::ptrdiff_t>( source ),
-                           volume.Voxels.begin() + static_cast<std::ptrdiff_t>( source + n * kBytesPerVoxel ),
+                           volume.Voxels.begin() +
+                                static_cast<std::ptrdiff_t>( source + static_cast<size_t>( n ) * kBytesPerVoxel ),
                            image.Pixels.begin() + static_cast<std::ptrdiff_t>( target ) );
             }
         }
@@ -176,7 +177,8 @@ namespace Desert::Assets
                 const size_t target = ( ( static_cast<size_t>( z ) * n + y ) * n ) * kBytesPerVoxel;
 
                 std::copy( pixels.begin() + static_cast<std::ptrdiff_t>( source ),
-                           pixels.begin() + static_cast<std::ptrdiff_t>( source + n * kBytesPerVoxel ),
+                           pixels.begin() +
+                                static_cast<std::ptrdiff_t>( source + static_cast<size_t>( n ) * kBytesPerVoxel ),
                            volume.Voxels.begin() + static_cast<std::ptrdiff_t>( target ) );
             }
         }
